Rename file to default_bytes_for_datatype.rs
