+++ src/distributed_planner/statistics/mod.rs
@@ -1,6 +1,6 @@
-mod calculate_bytes_per_row;
 mod calculate_row_stats;
 mod compute_per_node;
+mod default_bytes_for_datatype;
 
 pub(crate) use calculate_row_stats::calculate_row_stats;
 pub(crate) use compute_per_node::calculate_compute_complexity;
